refactor: npm update locked dep
