+++ lib/manager/npm/update/locked-dependency/common/parent-version.spec.ts
@@ -1,10 +1,10 @@
-import * as httpMock from '../../../../../test/http-mock';
-import { loadJsonFixture } from '../../../../../test/util';
+import * as httpMock from '../../../../../../test/http-mock';
+import { loadJsonFixture } from '../../../../../../test/util';
 import { findFirstParentVersion } from './parent-version';
 
 const expressJson = loadJsonFixture('express.json');
 
-describe('manager/npm/update/locked-dependency/parent-version', () => {
+describe('manager/npm/update/locked-dependency/common/parent-version', () => {
   describe('getLockedDependencies()', () => {
     it('finds indirect dependency', async () => {
       httpMock
