+++ lib/manager/npm/update/locked-dependency/package-lock/__fixtures__/accepts.json
