+++ lib/manager/npm/update/locked-dependency/package-lock/__fixtures__/package.json
