+++ lib/manager/npm/update/locked-dependency/package-lock/__fixtures__/type-is.json
