+++ lib/manager/npm/update/locked-dependency/package-lock/dep-constraints.spec.ts
@@ -1,12 +1,12 @@
-import { loadJsonFixture } from '../../../../../test/util';
+import { loadJsonFixture } from '../../../../../../test/util';
 import { findDepConstraints } from './dep-constraints';
 
-jest.mock('../../../../util/fs');
+jest.mock('../../../../../util/fs');
 
 const packageJson = loadJsonFixture('package.json');
 const packageLockJson = loadJsonFixture('package-lock.json');
 
-describe('manager/npm/update/locked-dependency/dep-constraints', () => {
+describe('manager/npm/update/locked-dependency/package-lock/dep-constraints', () => {
   describe('findDepConstraints()', () => {
     it('finds indirect dependency', () => {
       // FIXME: explicit assert condition
