+++ lib/manager/npm/update/locked-dependency/package-lock/get-locked.spec.ts
@@ -1,11 +1,11 @@
-import { loadJsonFixture } from '../../../../../test/util';
+import { loadJsonFixture } from '../../../../../../test/util';
 import { getLockedDependencies } from './get-locked';
 
-jest.mock('../../../../util/fs');
+jest.mock('../../../../../util/fs');
 
 const packageLockJson = loadJsonFixture('package-lock.json');
 
-describe('manager/npm/update/locked-dependency/get-locked', () => {
+describe('manager/npm/update/locked-dependency/package-lock/get-locked', () => {
   describe('getLockedDependencies()', () => {
     it('handles error', () => {
       expect(getLockedDependencies(null as any, 'some-dep', '1.0.0')).toEqual(
